feat: move everything to `modular_colpali.py`
